Refactor provider cache preparation and outcome reporting
